requirements.txt
apispec==6.7.1
blinker==1.9.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dateutil==2.9.0.post0
reportlab==4.2.5
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.1.3
